requirements.txt
streamlit>=1.28.0
pymilvus>=2.3.0
groq>=0.4.0
sentence-transformers>=2.2.0
tika>=2.6.0
python-dotenv>=1.0.0
httpx>=0.25.0
numpy>=1.24.0
pandas>=2.1.0
PyPDF2>=3.0.0
python-magic>=0.4.27
langchain>=0.1.0
langchain-community>=0.0.10 